// src/types.ts
export interface Product {
  id: string;
  name: string;
  price: number;
  description: string;
  image: string;
}

export interface Testimonial {
  id: string;
  name: string;
  content: string;
  image: string;
}

export interface Service {
  id: string;
  name: string;
  price: number;
  description: string;
  image: string;
}